graphql_gateway/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql[fastapi]
httpx # Async HTTP client
python-dotenv
python-jose[cryptography] # For JWT verification
pydantic[email] # Good practice
pydantic-settings